Enhance transition matrix structure by adding constants and types
This commit introduces a new constants file for transition matrix aggregation methods and updates the relevant imports in the TransitionMatrixPlot component. Additionally, a new types file is created to define the TransitionMatrixData interface, improving the organization and maintainability of the transition matrix module. The utility exports have also been adjusted for better clarity.
