+++ src/lib/plots/transition-matrix/const/index.ts
@@ -1,0 +1,9 @@
+/**
+ * Defines available aggregation methods for transition matrices
+ */
+export enum AggregationMethod {
+  SUM = 'sum',
+  PROBABILITY = 'probability',
+  DWELL_TIME = 'dwellTime',
+  SEGMENT_DWELL_TIME = 'segmentDwellTime',
+}
