+++ src/lib/plots/transition-matrix/index.ts
@@ -6,3 +6,6 @@ export type * from './types'
 
 // Export transition matrix utilities
 export * from './utils'
+
+// Export transition matrix constants
+export * from './const'
