+++ src/lib/plots/transition-matrix/types/index.ts
@@ -1,0 +1,10 @@
+import type { ExtendedInterpretedDataType } from '$lib/type/Data/InterpretedData/ExtendedInterpretedDataType'
+
+/**
+ * Interface representing an Transition Matrix with labels
+ */
+export interface TransitionMatrixData {
+  matrix: number[][]
+  aoiLabels: string[]
+  aoiList: ExtendedInterpretedDataType[]
+}
