+++ src/lib/plots/transition-matrix/utils/index.ts
@@ -1,1 +1,1 @@
-export * from './transitionMatrixTransformations'
+export * from './transformations'
